+++ src/DataFixtures/DemoTeamsFixtures.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\DataFixtures;
+
+use App\Entity\Team;
+use Doctrine\Bundle\FixturesBundle\Fixture;
+use Doctrine\Bundle\FixturesBundle\FixtureGroupInterface;
+use Doctrine\Common\DataFixtures\DependentFixtureInterface;
+use Doctrine\Persistence\ObjectManager;
+
+class DemoTeamsFixtures extends Fixture implements DependentFixtureInterface, FixtureGroupInterface
+{
+    public function load(ObjectManager $manager)
+    {
+        $team = new Team();
+        $team->setName('Green');
+        $manager->persist($team);
+        $team = new Team();
+        $team->setName('Blue');
+        $manager->persist($team);
+
+        $manager->flush();
+    }
+
+    public static function getGroups(): array
+    {
+        return ['demo'];
+    }
+
+    public function getDependencies(): array
+    {
+        return [
+            AppFixtures::class
+        ];
+    }
+}